+++ public/res/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Version 3.0.20200520 (dev)
+- Fixed options for event listener.
 - Fixed the value of data-mdc-dialog-button-default.
 
 ## Version 3.0.20200515 (dev)